OrderAhead Excel Add-in
Sample Excel add-in, for OrderAhead. Consumes the Orders CSV file and generates a couple of summaries, one by order status and another by quantity of items ordered. The purposes of the add-in are to demonstrate (1) my technical abilities with Excel and VBA, (2) my ability to apply design thinking, particularly the use of iteration for product development, and (3) the ability to convert business operations into a technical solution.